requirements.txt
openai
python-dotenv
pillow
requests_html
requests
azure-cosmos
pandas
azure-search-documents
azure-identity
yfinance
azure-functions
requests_html
pydantic
python-dotenv
grpcio-tools
starlette_gzip_request